+++ semgrep/semgrep/cli.py
@@ -8,9 +8,11 @@ import semgrep.config_resolver
 import semgrep.semgrep_main
 import semgrep.test
 from semgrep import __VERSION__
+from semgrep.bytesize import parse_size
 from semgrep.constants import DEFAULT_CONFIG_FILE
 from semgrep.constants import DEFAULT_MAX_CHARS_PER_LINE
 from semgrep.constants import DEFAULT_MAX_LINES_PER_FINDING
+from semgrep.constants import DEFAULT_MAX_TARGET_SIZE
 from semgrep.constants import DEFAULT_TIMEOUT
 from semgrep.constants import MAX_CHARS_FLAG_NAME
 from semgrep.constants import MAX_LINES_FLAG_NAME
@@ -19,11 +21,13 @@ from semgrep.constants import RCE_RULE_FLAG
 from semgrep.constants import SEMGREP_URL
 from semgrep.dump_ast import dump_parsed_ast
 from semgrep.error import SemgrepError
+from semgrep.metric_manager import metric_manager
 from semgrep.output import managed_output
 from semgrep.output import OutputSettings
 from semgrep.synthesize_patterns import synthesize_patterns
 from semgrep.target_manager import optional_stdin_target
 from semgrep.version import is_running_latest
+
 
 logger = logging.getLogger(__name__)
 try:
@@ -187,6 +191,18 @@ def cli() -> None:
         default=0,
         help=(
             "Maximum memory to use running a rule on a single file in MB. If set to 0 will not have memory limit. Defaults to 0."
+        ),
+    )
+
+    config.add_argument(
+        "--max-target-bytes",
+        type=parse_size,
+        default=DEFAULT_MAX_TARGET_SIZE,
+        help=(
+            "Maximum size for a file to be scanned by semgrep, e.g '1.5MB'. "
+            "Any input program larger than this will be ignored. "
+            "A zero or negative value disables this filter. "
+            f"Defaults to {DEFAULT_MAX_TARGET_SIZE} bytes."
         ),
     )
 
@@ -360,6 +376,21 @@ def cli() -> None:
         "--version", action="store_true", help="Show the version and exit."
     )
 
+    metric_group = parser.add_argument_group("config")
+    metric_ex = metric_group.add_mutually_exclusive_group()
+    metric_ex.add_argument(
+        "--enable-metrics",
+        action="store_true",
+        help="Opt-in to metrics. Defaults to what SEMGREP_SEND_METRICS envvar is set to",
+        default=os.environ.get("SEMGREP_SEND_METRICS"),
+    )
+    metric_ex.add_argument(
+        "--disable-metrics",
+        action="store_false",
+        help="Opt-out of metrics.",
+        dest="enable_metrics",
+    )
+
     parser.add_argument(
         "--force-color",
         action="store_true",
@@ -394,9 +425,15 @@ def cli() -> None:
 
     ### Parse and validate
     args = parser.parse_args()
+
     if args.version:
         print(__VERSION__)
         return
+
+    if args.enable_metrics:
+        metric_manager.enable()
+    else:
+        metric_manager.disable()
 
     if args.pattern and not args.lang:
         parser.error("-e/--pattern and -l/--lang must both be specified")
@@ -491,6 +528,7 @@ def cli() -> None:
                 jobs=args.jobs,
                 include=args.include,
                 exclude=args.exclude,
+                max_target_bytes=args.max_target_bytes,
                 strict=args.strict,
                 autofix=args.autofix,
                 dryrun=args.dryrun,